backend/requirements.txt
fastapi>=0.115
uvicorn>=0.30
sqlalchemy>=2.0
pandas>=2.2
numpy>=1.26
scipy>=1.13
scikit-learn>=1.5
matplotlib>=3.8
python-multipart>=0.0.9
pydantic-settings>=2.4
alembic>=1.13
requests>=2.31
statsmodels>=0.14
pyarrow>=18.0
PyMuPDF>=1.24
# Advanced Table and OCR (Requires Ghostscript/Poppler OS-level installs)
camelot-py[cv]>=0.11.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0

ruptures>=1.1.10
